Add security layer to check user access to place

+++ src/db/items.ts
@@ -23,8 +23,13 @@ export interface Item {
 
 export const getItemsForPlace = async (
   client: SupabaseClient,
-  placeId: number
+  placeId: number,
+  userId: number
 ): Promise<PostgrestResponse<Item>> => {
+  const hasAccess = await checkItemAccess(client, placeId, userId);
+  if (!hasAccess) {
+    throw new Error('User does not have access to this place');
+  }
   return client
     .from('pos_items')
     .select('*')
@@ -43,7 +48,10 @@ export const InsertItem = async (
   place_id: number,
   user_id: number
 ) => {
-  //get the bussiness id
+  const hasAccess = await checkItemAccess(client, place_id, user_id);
+  if (!hasAccess) {
+    throw new Error('User does not have access to this place');
+  }
   const business_id = await getUserBusinessId(client, user_id);
   let url = '';
   if (image) {
@@ -77,8 +85,15 @@ export const InsertItem = async (
 
 export const DeleteItem = async (
   client: SupabaseClient,
-  id: number
+  id: number,
+  place_id: number,
+  user_id: number
 ): Promise<PostgrestSingleResponse<Item>> => {
+  
+  const hasAccess = await checkItemAccess(client, place_id, user_id);
+  if (!hasAccess) {
+    throw new Error('User does not have access to this place');
+  }
   return client.from('pos_items').delete().eq('id', id).select().single();
 };
 
@@ -101,7 +116,16 @@ export const UpdateItem = async (
   item: Partial<Item>,
   image: File | null,
   user_id: number
-) => {
+): Promise<PostgrestSingleResponse<Item>> => {
+
+  if (!item.place_id) {
+    throw new Error('Place ID is required');
+  }
+
+  const hasAccess = await checkItemAccess(client, item.place_id, user_id);
+  if (!hasAccess) {
+    throw new Error('User does not have access to this place');
+  }
 
   const business_id = await getUserBusinessId(client, user_id);
   let url = item.image;
@@ -135,4 +159,17 @@ export const UpdateItemOrder = async (
     .eq('id', items.id)
     .select()
     .single();
+};
+
+export const checkItemAccess = async (
+  client: SupabaseClient,
+  placeId: number,
+  userId: number
+): Promise<boolean> => {
+
+  const data = await client.from('users').select('linked_business_id').eq('id', userId);
+  const business_id = data.data?.[0]?.linked_business_id;
+  const placesQuery = client.from('places').select('*').eq('business_id', business_id);
+  const place_ids = (await placesQuery).data?.map((place: any) => place.id) || [];
+  return place_ids.includes(placeId);
 };
